requirements.txt
tensorflow-gpu==2.11.0
opencv-python==4.5.5.64
keras==2.11.0
facenet==2.5.2
ultralytics==8.0.0
scikit-learn==1.2.0
numpy==1.24.0
matplotlib==3.7.0
pickle-mixin==1.0.2
keras-facenet